move `RankStatsWriter` class into rank_stats.py
