--- src/pheval/analyse/rank_stats_writer.py
@@ -1,59 +1,0 @@
-import csv
-from pathlib import Path
-
-from pheval.analyse.rank_stats import RankStats
-
-
-class RankStatsWriter:
-    """Write the rank stats for each run."""
-
-    def __init__(self, file: Path):
-        self.file = open(file, "w")
-        self.writer = csv.writer(self.file, delimiter="\t")
-        self.writer.writerow(
-            [
-                "results_directory_path",
-                "top",
-                "top3",
-                "top5",
-                "top10",
-                "found",
-                "total",
-                "mean_reciprocal_rank",
-                "percentage_top",
-                "percentage_top3",
-                "percentage_top5",
-                "percentage_top10",
-                "percentage_found",
-            ]
-        )
-
-    def write_row(self, directory: Path, rank_stats: RankStats) -> None:
-        """Write summary rank stats row for run."""
-        try:
-            self.writer.writerow(
-                [
-                    directory,
-                    rank_stats.top,
-                    rank_stats.top3,
-                    rank_stats.top5,
-                    rank_stats.top10,
-                    rank_stats.found,
-                    rank_stats.total,
-                    rank_stats.mean_reciprocal_rank(),
-                    rank_stats.percentage_top(),
-                    rank_stats.percentage_top3(),
-                    rank_stats.percentage_top5(),
-                    rank_stats.percentage_top10(),
-                    rank_stats.percentage_found(),
-                ]
-            )
-        except IOError:
-            print("Error writing ", self.file)
-
-    def close(self) -> None:
-        """Close file."""
-        try:
-            self.file.close()
-        except IOError:
-            print("Error closing ", self.file)
